add menu component
replace burger menu to component, refactor logic
