+++ src/components/Header/HeaderContent/HeaderMenu.module.scss
@@ -5,6 +5,7 @@
     position: fixed;
     top: 696px;
     z-index: 100;
+    // margin-left: -432px;
   }
 }
