+++ src/components/Music/MusicComponent.jsx
@@ -10,6 +10,7 @@ import ListMusic from './ListMusic/ListMusic';
 import ButtonShowMoreMusic from './ButtonShowMoreMusic/ButtonShowMoreMusic';
 
 import scss from './MusicComponent.module.scss';
+import Menu from 'components/Menu/Menu';
 
 function MusicComponent() {
   const { data } = useFetch('music');
@@ -27,7 +28,10 @@ function MusicComponent() {
   return (
     <section className={scss.musicComponent}>
       <div className="container">
-        <HeaderMenu page="Music" text="Back to Home" />
+        <Menu />
+        <div className={scss.headermenu}>
+          <HeaderMenu page="Music" text="Back to Home" />
+        </div>
         {data && data.title[currentLanguage] && (
           <ChapterMusic chapterMusic={data.title[currentLanguage]} />
         )}
